backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.31.0
httpx==0.24.1
APScheduler==3.10.4
pytest==7.4.0
pytest-cov==4.1.0
Werkzeug==2.3.6
